requirements.txt
pyobjc
PyQt5
pynput
pystray
